Add `deleteAll` query method to `RoleRepository` and refactor `RoleServiceImpl`
Check if roles exist before creating or deleting roles.

+++ src/main/java/dev/naiarievilo/todoapp/roles/RoleServiceImpl.java
@@ -49,19 +49,22 @@ public class RoleServiceImpl implements RoleService {
     @Override
     @Transactional
     public void createRole(Roles role) {
-        String roleName = role.name();
-        String description = role.description();
+        if (roleExists(role)) {
+            throw new RoleAlreadyExistsException(role.name());
+        }
 
         Role newRole = new Role();
-        newRole.setName(roleName);
-        newRole.setDescription(description);
-
+        newRole.setName(role.name());
+        newRole.setDescription(role.description());
         roleRepository.persist(newRole);
     }
 
     @Override
     @Transactional
     public void deleteRole(Roles role) {
+        if (!roleExists(role)) {
+            throw new RoleNotFoundException(role.name());
+        }
         roleRepository.deleteByName(role.name());
     }
